MBB increase size uses different metric. Changed demo run structure

+++ main.py
@@ -23,7 +23,7 @@ def delete_saved_rtree():
 # rtree 1102B
 # database 81B
 
-if __name__ == '__main__':
+def create_new_tree(delete_after: bool = True):
     print("hello, friend")
     # print(cpu_count())
     tree = RTree()
@@ -31,17 +31,24 @@ if __name__ == '__main__':
     x, y = -1, -1
     total_insert_count = 0
     try:
-        random_x = list(range(0, 500))
-        random.shuffle(random_x)
-        random_x = random_x[0:140]
-        for x in random_x:
-            random_y = list(range(0, 500))
-            random.shuffle(random_y)
-            random_y = random_y[0:140]
-            for y in random_y:
-                # print(f"x: {x}; y: {y}")
-                total_insert_count += 1
-                tree.insert_entry(DatabaseEntry(coordinates=[x, y], data=f"This is generated x: {x}; y: {y}"))
+        # random_x = list(range(0, 500))
+        # random.shuffle(random_x)
+        # random_x = random_x[0:6]
+        # for x in random_x:
+        #     random_y = list(range(0, 500))
+        #     random.shuffle(random_y)
+        #     random_y = random_y[0:6]
+        #     for y in random_y:
+        #         # print(f"x: {x}; y: {y}")
+        #         total_insert_count += 1
+        #         tree.insert_entry(DatabaseEntry(coordinates=[x, y], data=f"This is generated x: {x}; y: {y}"))
+
+        for _ in range(0, 100):
+            x = random.randint(100, 200)
+            y = random.randint(100, 200)
+
+            total_insert_count += 1
+            tree.insert_entry(DatabaseEntry(coordinates=[x, y], data=f"This is generated x: {x}; y: {y}"))
 
         print("\n==========================")
         for found_node in tree.search_rectangle(coordinates_min=[0, 0], coordinates_max=[20, 20]):
@@ -50,28 +57,34 @@ if __name__ == '__main__':
         for found_node in tree.search_k_nearest_neighbours(4, coordinates=[0, 4]):
             print(found_node)
 
-        # visualize(tree, show_mbbs_only=False)  # False)
+        visualize(tree, show_mbbs_only=False)  # False)
     except Exception as e:
         print(f"x: {x}; y: {y}")
         print(f"total_insert_count {total_insert_count}")
         traceback.print_exc()
 
     print(f"total_insert_count {total_insert_count}")
-    tree.tree_handler.file.close()
-    tree.database.file.close()
-    delete_saved_rtree()
+    # tree.tree_handler.file.close()
+    # tree.database.file.close()
+    del tree
+    if delete_after:
+        delete_saved_rtree()
 
-    # if not load_from_files:
-    #     new_entry = DatabaseEntry([0 for coord in range(self.dimensions)],
-    #                               data="this is some data")
-    #     new_entry_id = self.database.create(new_entry)
-    #     root_node = self.tree_handler.get_node(self.root_id)
-    #     if root_node is None:
-    #         raise Exception("root node not found")
-    #
-    #     root_node.insert_entry_from_entry(new_entry_id, new_entry)
-    #
-    #     self.tree_handler.update_node(self.root_id, root_node)
+
+def load_tree(delete_after: bool = True):
+    try:
+        tree = RTree()
+        visualize(tree, show_mbbs_only=False)
+    except Exception:
+        traceback.print_exc()
+    del tree
+    if delete_after:
+        delete_saved_rtree()
+
+
+if __name__ == '__main__':
+    create_new_tree(delete_after=True)
+    # load_tree(delete_after = False)
 
 """
 Non-leaf node:
